ASSIGNED - bug 71: Biblefox 1.0 http://dev.may26.org/bugzilla/show_bug.cgi?id=71
HACK: is_main_blog isn't enough - when a new blog is created, it needs to be activated, which for instance:
The url http://demo.biblefox.com/wp-activate.php would go to blog #1, but the domain is demo.biblefox.com, so is_main_blog() would return false even though we are using blog #1. This was leading to main_blog.php not being loaded even though we were loading the main blog theme which needs it.